META-420: Rename push commands to apply
