+++ lib/application/applicationsRouter.js
@@ -30,8 +30,8 @@ const cmdDefinitions = [
     requirements: [CommandRequirement.AUTH]
   },
   {
-    command: 'push',
-    desc: 'Update an application',
+    command: 'apply',
+    desc: 'Apply an application configuration file',
     builder: (commandsManager) => {
       commandsManager
         .option('file', { describe: 'Path to an application configuration file', required: true })
