+++ lib/environment/environmentsRouter.js
@@ -30,8 +30,8 @@ const cmdDefinitions = [
     requirements: [CommandRequirement.AUTH]
   },
   {
-    command: 'push',
-    desc: 'Update an environment',
+    command: 'apply',
+    desc: 'Apply an environment configuration file',
     builder: (commandsManager) => {
       commandsManager
         .option('file', { describe: 'Path to an environment configuration file', required: true })
