+++ lib/organization/organizationsRouter.js
@@ -17,8 +17,8 @@ const { CommandRequirement, ExportOptions, ExportOptionsNames, Namespace, OrgOpt
 
 const cmdDefinitions = [
   {
-    command: 'push',
-    desc: 'Update an organization',
+    command: 'apply',
+    desc: 'Apply an organization configuration file',
     builder: (commandsManager) => {
       commandsManager
         .option('file', { describe: 'Path to an organization configuration file', required: true })
