+++ lib/service/servicesRouter.js
@@ -32,8 +32,8 @@ const cmdDefinitions = [
     requirements: [CommandRequirement.AUTH]
   },
   {
-    command: 'push',
-    desc: 'Update a service',
+    command: 'apply',
+    desc: 'Apply a service configuration file',
     builder: (commandsManager) => {
       commandsManager
         .option('file', { describe: 'Path to a service configuration file', required: true })
